requirements.txt
fastdtw==0.3.4
jamo==0.4.1
librosa==0.10.1
matplotlib==3.9.0
numpy==1.26.2
omegaconf==2.3.0
openai-whisper==20231117
openphonemizer==0.1.2
pycwt==0.4.0b0
pydantic==2.8.2
pysptk==0.2.2
pyworld==0.3.4
pyyaml==6.0.1
scipy==1.11.1
speechbrain==0.5.16
streamlit==1.23.1
streamlit-option-menu==0.3.6
tabulate==0.9.0
torch==2.0.1
torchaudio==2.0.2
torchcrepe==0.0.23
transformers==4.41.2